apps/core/resource_api.py
"""Resource Registry configuration for DAB."""

from ansible_base.feature_flags.models import AAPFlag
from ansible_base.resource_registry.registry import (
    ParentResource,
    ResourceConfig,
    ServiceAPIConfig,
    SharedResource,
)
from ansible_base.resource_registry.shared_types import FeatureFlagType, UserType

from apps.core.models import Organization, Team, User


class APIConfig(ServiceAPIConfig):
    """API configuration for the resource registry."""

    service_type = "inventory_service"


RESOURCE_LIST = [
    ResourceConfig(
        Organization,
        shared_resource=SharedResource(
            serializer=None,
            is_provider=True,
        ),
    ),
    ResourceConfig(
        Team,
        shared_resource=SharedResource(
            serializer=None,
            is_provider=True,
        ),
        parent_resources=[
            ParentResource(model=Organization, field_name="organization"),
        ],
    ),
    ResourceConfig(
        User,
        shared_resource=SharedResource(
            serializer=UserType,
            is_provider=False,
        ),
        name_field="username",
    ),
    ResourceConfig(
        AAPFlag,
        shared_resource=SharedResource(serializer=FeatureFlagType, is_provider=False),
    ),
]
