pyproject.toml
[project]
name = "inventory_service"
version = "0.1.0"
description = "description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.2.7",
    "psycopg[binary]>=3.3.1",
    "django-ansible-base[rest_filters,jwt_consumer,resource_registry,rbac,feature_flags,api_documentation]",
    "dispatcherd[pg_notify] @ git+https://github.com/ansible/dispatcherd",
    "django-filter>=25.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.4.2",
    "pytest-django>=4.10.0",
    "poethepoet>=0.37.0",
    "ruff>=0.14.3",
    "ty>=0.0.1a25",
    "pdoc>=16.0.0",
    "django-extensions>=4.1",
    "ipython>=9.7.0",
    "ipdb>=0.13.13",
    "django-extensions>=4.1",
    "ipython>=9.7.0",
    "ipdb>=0.13.13",
    "django-debug-toolbar>=6.1.0",
]
doc = [
    "pdoc>=16.0.0",
]
test = [
    "pytest>=8.4.2",
    "poethepoet>=0.37.0",
    "ruff>=0.14.3",
    "ty>=0.0.1a25",
]

[tool.ruff]
line-length = 120
exclude = [
  "__pycache__",
  ".venv",
  "build",
  "dist",
  ".git",
  "migrations",
  "tests",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "N", "I"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ty.src]
exclude = [
    "tests",
    "apps/**/tests",
    "migrations",
]

[tool.ty.rules]
possibly-missing-implicit-call = "ignore"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vs -ra -q"
testpaths = ["inventory_service/tests", "apps/**/tests"]
DJANGO_SETTINGS_MODULE = "inventory_service.settings"

[tool.poe]
executor.type  = "uv"

[tool.poe.tasks]
lint = "ruff check ."
type_check = "ty check"
format = "ruff format ."
unit-test = "manage.py test -v 2"
check = ["format", "lint", "unit-test"]
render-docs = "pdoc -t docs/templates --mermaid --docformat markdown --docformat google inventory_service apps -o html"
serve-docs = "pdoc -t docs/templates --mermaid --docformat markdown --docformat google inventory_service apps"

[tool.poe.tasks.clean]
help = "Clean local caches for mypy, pytest, python"
script = "poethepoet.scripts:rm('.mypy_cache', '.pytest_cache', './**/__pycache__', './**/*.pyc')"

[tool.poe.tasks.validate]
shell = """
python -c "
import yaml
with open('.copier-answers.yml') as f:
    d = yaml.safe_load(f)
    branch = d.get('src_branch', '')
    src = d.get('_src_path', '')
    from pathlib import Path
    if Path(src).is_dir():
        print(f'uvx --refresh --from {src} platform-service-framework validate')
    else:
        print(f'uvx --refresh git+{src}@{branch} validate')
" | sh
"""

[tool.poe.tasks.update]
help = "Update the project from template (use --core to also update core templates)"
args = [
  { name = "core", help = "Also update core templates", type = "boolean" }
]
shell = """
python -c "
import yaml
import os
with open('.copier-answers.yml') as f:
    d = yaml.safe_load(f)
    branch = d.get('src_branch', '')
    src = d.get('_src_path', '')
    core = '$core'
    core_flag = ' --core' if core == 'True' else ''
    from pathlib import Path
    if Path(src).is_dir():
        print(f'uvx --refresh --from {src} platform-service-framework update{core_flag}')
    else:
        print(f'uvx --refresh git+{src}@{branch} update{core_flag}')
" | sh
"""

[tool.uv.sources]
django-ansible-base = { git = "https://github.com/ansible/django-ansible-base", rev = "devel" }
dispatcherd = { git = "https://github.com/ansible/dispatcherd" }

[tool.setuptools.packages.find]
where = ["."]
include = ["inventory_service*", "apps*"]
exclude = ["tests*"]
